Make changes in response to PR comments
